feat: added support for sb3 project conversion
