+++ src/components/TopBar/FileDropDown/index.ts
@@ -1,0 +1,1 @@
+export * from './component';
